Merged everything relevant from govCMS/govcms-project/ into the base Platform.sh Drupal9 template
This was done with reference to the GovCMS Docs about building the project with composer https://github.com/govCMS/GovCMS#via-composer-create-project
so we add the contents of: https://github.com/govCMS/govcms-project/blob/2.x/composer.json
Includes a couple of extra repository locations and allowed plugins, but basically we should be able to let composer do everything.

This should just include the govCMs/GovCMS profile as a Drupal component, and take care of the remaining dependencies from there.

https://github.com/govCMS/GovCMS
The profile itself declares the need for drush and drupal core (and the profile may be pinned to a specific version), so we don't define there here.

At this date, we got GovCMS 2.23.0 and Drupal 9.4.7, (as defined and pinned by GovCMS)
https://packagist.org/packages/govcms/govcms#2.23.0

To rebuild properly, needed to nuke composer.lock before running install.